# Improved ANSI passthrough.