--- a/tests/integration_tests.rs
+++ b/tests/integration_tests.rs
@@ -1138,3 +1138,22 @@ fn grid_for_file_without_newline() {
         )
         .stderr("");
 }
+
+// Ensure that ANSI passthrough is emitted properly for both wrapping and non-wrapping printer.
+#[test]
+fn ansi_passthrough_emit() {
+    for wrapping in vec!["never", "character"] {
+        bat()
+            .arg("--paging=never")
+            .arg("--color=never")
+            .arg("--terminal-width=80")
+            .arg(format!("--wrap={}", wrapping))
+            .arg("--decorations=always")
+            .arg("--style=plain")
+            .write_stdin("\x1B[33mColor\nColor \x1B[m\nPlain\n")
+            .assert()
+            .success()
+            .stdout("\x1B[33m\x1B[33mColor\n\x1B[33mColor \x1B[m\nPlain\n")
+            .stderr("");
+    }
+}