This pull request overhauls how `bat` interprets and re-emits ANSI escape sequences.
Prior to these changes, `bat` used a naiive heuristic that re-emits every encountered sequence on a new line. The encountered sequences are only cleared whenever the exact literal `\x1B[0m` is encountered.

In 60aad681f358fc7285ea72c77faa8782a060eab1, I added a simple ANSI sequence parser that only tracks the most recently-encountered ANSI SGR (color, style, etc.) sequence for each attribute. I haven't benchmarked the difference in CPU usage, but it significantly increases the performance of the pager (the cause of https://github.com/sharkdp/bat/issues/1481).

**Memory usage: (Top: before, Bottom: after)**
![image](https://user-images.githubusercontent.com/32112321/112074268-2e38cf00-8b33-11eb-8101-039db7bfdfb7.png)

In 5aa94e8765e992cf669fcbb71a54ae8f5a57554c, I fixed ANSI passthrough support when wrapping is disabled. The non-wrapping branch didn't have code to handle ANSI, which led to inconsistencies across lines.

**Example: `printf '\x1B[33mYellow\nShould be yellow.\x1B[m' | bat --wrap=never`**
![image](https://user-images.githubusercontent.com/32112321/112080126-a3f66800-8b3e-11eb-974a-530d00214900.png)

---

I did my best to manually check for any regressions, but I may have missed some. If anyone finds any issues with how ANSI is handled now, please let me know.

---

Additionally, if/when https://github.com/mitsuhiko/console/pull/95 is merged, we should update it. Without the fix, strings such as `\x1B(0lqk` will be incorrectly interpreted as [`\x1B(0l`, `qk`] instead of [`\x1B(0`, `lqk`].

**Edit by @Enselic:** We included the fix in bat a month ago in https://github.com/sharkdp/bat/pull/1934 :tada: