@eth-p 2021-03-28
> It would be great to see actual (before and after) benchmark results.

I unfortunately don't have a machine that's reliable enough to accurately benchmark `bat`'s wall time. It seems my laptop can introduce noise (thermal throttling? background CPU usage?) that may vary results by up to 40% per run, and I don't want to provide any unreliable benchmarks.

That being said... I can still quantify the difference in how many bytes `bat` is writing to stdout or the pager. In some cases, it's a small but meaningful 9% difference.

**Using the pre-highlighted Rust syntax test:**

<table>
<tr><th>Cols</th><th>Syntax</th><th>Style</th><th>Picture</th><th>OLD</th><th>NEW</th><th>Diff%</th></tr>
<tr><td>80</td><td>text</td><td>Plain</td><td><img src="https://user-images.githubusercontent.com/32112321/112740322-e77f1680-8f30-11eb-80f5-dd565f50a4dc.png"/></td><td>92016</td><td>83801</td><td>91.07%</td></tr>
<tr><td>120</td><td>Rust</td><td>Plain</td><td><img src="https://user-images.githubusercontent.com/32112321/112740354-2d3bdf00-8f31-11eb-9ebf-50d80e1fd472.png"/></td><td>99073</td><td>91131</td><td>91.98%</td></tr>
<tr><td>120</td><td>Rust</td><td>Full</td><td><img src="https://user-images.githubusercontent.com/32112321/112740372-565c6f80-8f31-11eb-8818-fbb3006acbba.png"/></td><td>692732</td><td>634222</td><td>91.55%</td></tr>
</table>

**And in more extreme cases such as the output from the command in #1481:**

<table>
<tr><th>Cols</th><th>Syntax</th><th>Style</th><th>Picture</th><th>OLD</th><th>NEW</th><th>Diff%</th></tr>
<tr><td>120</td><td>text</td><td>Full</td><td><img src="https://user-images.githubusercontent.com/32112321/112740717-38dcd500-8f34-11eb-97c7-f940ebab18f0.png"/></td><td>908560976</td><td>1082842</td><td>0.1191%</td></tr>
</table>

As far as the changes in 5aa94e8765e992cf669fcbb71a54ae8f5a57554c are concerned though, they **will** have negative impact on unhighlighted text with `--wrap=never`. Prior to that commit, `--wrap=never` was faster than it should have been, since ANSI sequences were not actually being interpreted.

> Maybe with some of the "highlighted" files in tests/syntax-tests (which doesn't seem to work currently)?

They do work with `-l txt`. I suspect the reason they don't work is because Syntect is highlighting parts of the ANSI color sequences as their own highlight blocks, and `bat` doesn't try to look for sequences across blocks.

A smarter implementation would likey be for `bat` to just disable all highlighting when it encounters its first ANSI escape sequence, but I feel like that would be a breaking change that requires some discussion around it first.