@sharkdp 2021-04-03
> I unfortunately don't have a machine that's reliable enough to accurately benchmark `bat`'s wall time. It seems my laptop can introduce noise (thermal throttling? background CPU usage?) that may vary results by up to 40% per run, and I don't want to provide any unreliable benchmarks.

I wrote [`hyperfine`](https://github.com/sharkdp/hyperfine) to be able to deal with effects like these when benchmarking command line programs. I'm not asking you to use it, but if you are interested, you could give it a try :smile: 

```
hyperfine [options] './bat-old …' './bat-new'
```

- There is a builtin outlier detection that will tell you if background processes completely break your benchmark. Obviously, you should still always try to run benchmarks on a quiet PC. Shuting down things like spotify, dropbox, browsers, … can have a huge impact.
- Hyperfine performs a series of benchmarks and performs a statistical analysis. You will get an estimate for the noise in your benchmark results. If you are still unsure, you can run a [t-test](https://github.com/sharkdp/hyperfine/blob/master/scripts/welch_ttest.py) on the results that will tell you if the speedup/slowdown is statistically significant.
- There are options to use warmup runs before the actual benchmark. This can help with disk caching effects (which are likely to play a role here as well) and possibly even with thermal throttling.

Concerning the last point (thermal throttling): On Linux, there is a way to temporarily disable any [frequency scaling](https://wiki.archlinux.org/index.php/CPU_frequency_scaling#Scaling_governors), which can help with benchmarking.