@eth-p 2021-04-03
@sharkdp That worked really well, even on my Mac.

Okay, so... as expected, this pull request does introduce some cost at runtime. In the worst case (e.g. #1481), it takes 27% longer when not printing to a terminal:

![image](https://user-images.githubusercontent.com/32112321/113490575-ce6dec80-947f-11eb-9169-deff0c53570e.png)

And in the best case (no ANSI), it doesn't really change anything:

![image](https://user-images.githubusercontent.com/32112321/113490619-27d61b80-9480-11eb-830e-5e56da119d8d.png)


Now when it actually prints something to the screen... let's just say that the old one couldn't be benchmarked.

`hyperfine -r10 --show-output './bat-old -ltxt --decorations=always --color=always --paging=never lots-of-ansi.txt'`

![image](https://user-images.githubusercontent.com/32112321/113490843-9798d600-9481-11eb-8619-7acddf5553e1.png)

The fixed version worked *a lot* better:

![image](https://user-images.githubusercontent.com/32112321/113491089-f448c080-9482-11eb-881a-52a2200685d5.png)