@Enselic 2021-11-19
@ImportTaste Does passing `--color=never --wrap=never` preserve colors? See https://github.com/sharkdp/bat#garbled-output (you did not miss that, it was added minutes ago)