@Enselic 2021-12-08
I merged the code with origin/master and fixed some lints.

I have confirmed that the added integration test fails on master. In other words, it tests that the fix in this PR is necessary and works. Which is a great test to have.

I do think the code could use some de-duplication, but that was the case even before this PR, so I don't think we need to do that as part of this PR.

Unless I'm mistaken, I think that #1976 allows us to do comprehensive benchmarking of this PR. Here are the results on my low-end desktop:

* The `grep-output-ansi-sequences.txt` test is much faster, from 912 ms to 177 ms with `--wrap=character`.
* Performance for other files with `--wrap=character` is unchanged.
* As already mentioned by Ethan, performance does worsen quite a bit with `--wrap=never` for `grep-output-ansi-sequences.txt`, going from 129 ms to 157 ms. But, as explained, that's because we cheated before.
* For other files with `--wrap=never` (that does not contain ANSI escape sequences) there is very little change in performance.

IMHO we should go ahead and merge this PR now, because it is an overall improvement. And we can always keep working on the code.

Here is the full benchmark comparison if you want to look:
![master-vs-this-pr](https://user-images.githubusercontent.com/115040/145198438-5a0023f1-1127-4459-b992-3874f82d81bd.png)